requirements.txt
fastapi
uvicorn
pillow
numpy
torch==2.0.1+cpu
torchvision==0.15.2+cpu
-f https://download.pytorch.org/whl/torch_stable.html
easyocr
python-multipart
